Distribute display text tests into per-spec test files
Move GetDisplayText tests from DisplayTextTests.cs into the
existing test class for each spec. Move CnctConfig.ExecuteAsync
display/label tests into CnctConfigTagFilteringTests.cs.
Delete DisplayTextTests.cs.
