+++ Cnct/Cnct.Core.Tests/CloneGitRepositoryTaskSpecificationTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.IO;
 using System.IO.Abstractions.TestingHelpers;
 using System.Threading.Tasks;
@@ -219,6 +220,47 @@ namespace Cnct.Core.Tests
             logger.Verify(l => l.LogWarning(It.Is<string>(s => s.Contains(dest))), Times.Once);
             mockRunner.Verify(r => r.CloneAsync(It.IsAny<string>(), It.IsAny<string>()), Times.Never);
             mockRunner.Verify(r => r.PullAsync(It.IsAny<string>()), Times.Never);
+        }
+
+        [Fact]
+        public void GetDisplayText_SingleRepo_ReturnsActionTypeAndUri()
+        {
+            var spec = new CloneGitRepositoryTaskSpecification
+            {
+                Repos = new ReadOnlyDictionary<string, string>(
+                    new Dictionary<string, string>
+                    {
+                        ["https://github.com/user/repo"] = "~/repos/repo",
+                    }),
+            };
+
+            Assert.Equal("cloneGitRepository: https://github.com/user/repo", spec.GetDisplayText());
+        }
+
+        [Fact]
+        public void GetDisplayText_MultipleRepos_ReturnsAllUris()
+        {
+            var spec = new CloneGitRepositoryTaskSpecification
+            {
+                Repos = new ReadOnlyDictionary<string, string>(
+                    new Dictionary<string, string>
+                    {
+                        ["https://github.com/user/repo1"] = "~/repos/repo1",
+                        ["https://github.com/user/repo2"] = "~/repos/repo2",
+                    }),
+            };
+
+            Assert.Equal(
+                "cloneGitRepository: https://github.com/user/repo1, https://github.com/user/repo2",
+                spec.GetDisplayText());
+        }
+
+        [Fact]
+        public void GetDisplayText_NullRepos_ReturnsActionType()
+        {
+            var spec = new CloneGitRepositoryTaskSpecification();
+
+            Assert.Equal("cloneGitRepository", spec.GetDisplayText());
         }
     }
 }
